+++ execution_engine/src/storage/global_state/scratch.rs
@@ -48,7 +48,7 @@ impl Cache {
     }
 
     /// Consumes self and returns only written values as values that were only read must be filtered
-    /// out to prevent unecessary writes.
+    /// out to prevent unnecessary writes.
     fn into_dirty_writes(self) -> HashMap<Key, StoredValue> {
         self.cached_values
             .into_iter()