+++ node/src/effect/requests.rs
@@ -31,7 +31,7 @@ use casper_execution_engine::{
 use casper_hashing::Digest;
 use casper_types::{
     bytesrepr::Bytes, system::auction::EraValidators, EraId, ExecutionResult, Key, ProtocolVersion,
-    PublicKey, Transfer, URef,
+    PublicKey, TimeDiff, Transfer, URef,
 };
 
 use crate::{
@@ -51,7 +51,7 @@ use crate::{
         AvailableBlockRange, Block, BlockHash, BlockHeader, BlockHeaderWithMetadata, BlockPayload,
         BlockSignatures, BlockWithMetadata, Chainspec, ChainspecInfo, ChainspecRawBytes, Deploy,
         DeployHash, DeployMetadata, DeployWithFinalizedApprovals, FinalizedApprovals,
-        FinalizedBlock, Item, NodeId, StatusFeed, TimeDiff,
+        FinalizedBlock, Item, NodeId, StatusFeed,
     },
     utils::{DisplayIter, Source},
 };
@@ -290,6 +290,16 @@ pub(crate) enum StorageRequest {
         /// local storage.
         responder: Responder<Option<BlockHeader>>,
     },
+    GetBlockHeaderByHeight {
+        /// Height of block to get header of.
+        block_height: u64,
+        /// Flag indicating whether storage should check the block availability before trying to
+        /// retrieve it.
+        only_from_available_block_range: bool,
+        /// Responder to call with the result.  Returns `None` is the block header doesn't exist in
+        /// local storage.
+        responder: Responder<Option<BlockHeader>>,
+    },
     /// Checks if a block header at the given height exists in storage.
     CheckBlockHeaderExistence {
         /// Height of the block to check.
@@ -431,11 +441,6 @@ pub(crate) enum StorageRequest {
         /// Responder to call with the result.
         responder: Responder<AvailableBlockRange>,
     },
-    /// Get the lowest block of the available range.
-    GetLowestAvailableBlock {
-        /// Responder to call with the block if it exists.
-        responder: Responder<Option<Block>>,
-    },
     /// Store a set of finalized approvals for a specific deploy.
     StoreFinalizedApprovals {
         /// The deploy hash to store the finalized approvals for.
@@ -464,6 +469,9 @@ impl Display for StorageRequest {
             }
             StorageRequest::GetBlockHeader { block_hash, .. } => {
                 write!(formatter, "get {}", block_hash)
+            }
+            StorageRequest::GetBlockHeaderByHeight { block_height, .. } => {
+                write!(formatter, "get header for height {}", block_height)
             }
             StorageRequest::CheckBlockHeaderExistence { block_height, .. } => {
                 write!(formatter, "check existence {}", block_height)
@@ -543,9 +551,6 @@ impl Display for StorageRequest {
             }
             StorageRequest::GetAvailableBlockRange { .. } => {
                 write!(formatter, "get available block range",)
-            }
-            StorageRequest::GetLowestAvailableBlock { .. } => {
-                write!(formatter, "get lowest available block",)
             }
             StorageRequest::StoreFinalizedApprovals { deploy_hash, .. } => {
                 write!(formatter, "finalized approvals for deploy {}", deploy_hash)