+++ node/src/types/chainspec/core_config.rs
@@ -4,11 +4,12 @@ use num::rational::Ratio;
 use rand::Rng;
 use serde::{Deserialize, Serialize};
 
-use casper_types::bytesrepr::{self, FromBytes, ToBytes};
-
 #[cfg(test)]
-use crate::testing::TestRng;
-use crate::types::TimeDiff;
+use casper_types::testing::TestRng;
+use casper_types::{
+    bytesrepr::{self, FromBytes, ToBytes},
+    TimeDiff,
+};
 
 #[derive(Copy, Clone, DataSize, PartialEq, Eq, Serialize, Deserialize, Debug)]
 // Disallow unknown fields to ensure config files and command-line overrides contain valid keys.