+++ types/src/system/auction/era_info.rs
@@ -251,7 +251,7 @@ impl CLTyped for EraInfo {
 }
 
 /// Generators for [`SeigniorageAllocation`] and [`EraInfo`]
-#[cfg(any(feature = "gens", test))]
+#[cfg(any(feature = "testing", test))]
 pub mod gens {
     use proptest::{
         collection::{self, SizeRange},