+++ execution_engine_testing/test_support/CHANGELOG.md
@@ -14,14 +14,18 @@ All notable changes to this project will be documented in this file.  The format
 ## [Unreleased]
 
 ### Added
-* Add `WasmTestBuilder::get_execution_journals` method for returning execution journals for all test runs.
+* Added `WasmTestBuilder::get_execution_journals` method for returning execution journals for all test runs.
+* Added support to load values from a given Chainspec.
+* Added static and constants that represent Casper-mainnet chainspec values. These values will change as new ProtocolVersions are added. The current values reflect ones used in the 1.5.0 ProtocolVersion.
+* Added `WasmTestBuilder::advance_era`, `WasmTestBuilder::advance_eras_by`, and `WasmTestBuilder::advance_eras_by_default_auction_delay` to advance chain and run auction contract in test environment.
 
 ### Changed
 * Extract some auction and transfer bench code to test_support in support of new `disk_use` binary.
 * `WasmTestBuilder::get_transforms` is deprecated in favor of `WasmTestBuilder::get_execution_journals`.
 * `deploy_hash` field is now defaulted to a random value rather than zeros in `DeployItemBuilder`.
 
-
+### Deprecated
+* Deprecated the `DEFAULT_GENESIS_REQUEST` in favor of `PRODUCTION_GENESIS_REQUEST`.
 
 ## 2.1.0
 
@@ -65,7 +69,7 @@ All notable changes to this project will be documented in this file.  The format
     * `StepRequestBuilder` for building a `StepRequest` (generally only used by the execution engine itself)
     * `UpgradeRequestBuilder` for building an `UpgradeRequest` (generally only used by the execution engine itself)
 * Provide `InMemoryWasmTestBuilder` which will be suitable in most cases for testing a smart contract
-* Provide `LmdbWasmTestBuilder` can be used where global state needs to be persisted after execution of a smart contract
+* Provide `DbWasmTestBuilder` can be used where global state needs to be persisted after execution of a smart contract
 * Provide several helper functions in `utils` module
 * Provide several default consts and statics useful across many test scenarios